Struct rustcv::core::Mat

pub struct Mat { /* fields omitted */ }

The class Mat represents an n-dimensional dense numerical single-channel or multi-channel array. It can be used to store real or complex-valued vectors and matrices, grayscale or color images, voxel volumes, vector fields, point clouds, tensors, histograms

Methods

impl Mat

pub fn new() -> Mat

Creates an empty Mat struct.

pub fn new_with_size(rows: i32, cols: i32, t: CvType) -> Mat

Creates an empty Mat struct with size.

pub fn new_from_scalar(s: Scalar, t: CvType) -> Mat

Creates an empty Mat struct with a constant scalar.

pub fn new_from_bytes(rows: i32, cols: i32, t: CvType, buf: &mut [i8]) -> Mat

Creates an empty Mat struct from buffer.

pub fn empty(&self) -> bool

Determines if the Mat is empty or not.

pub fn copy_to(&self, dst: &mut Mat)

Copies Mat into destination Mat.

For further details, please see OpenCV documentation

pub fn copy_to_with_mask(&self, dst: &mut Mat, mask: &Mat)

Copies Mat into destination Mat, with a mask.

pub fn convert_to(&self, dst: &mut Mat, t: CvType)

Converts Mat into destination Mat.

For further details, please see OpenCV documentation

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn to_bytes(&self) -> Vec<u8>

Copies the underlying Mat data to a byte array.

For further details, please see OpenCV documentation

pub fn region(&self, roi: Rect) -> Mat

Returns a new Mat that points to a region of this Mat. Changes made to the region Mat will affect the original Mat, since they are pointers to the underlying OpenCV Mat object.

pub fn reshape(&self, channel: i32, rows: i32) -> Mat

Changes the shape and/or the number of channels of a 2D matrix without copying the data. The method makes a new matrix header for the internal data.

pub fn convert_fp16(&self) -> Mat

Converts a Mat to half-precision floating point.

This function converts FP32 (single precision floating point) from/to FP16 (half precision floating point). The input array has to have type of CV_32F or CV_16S to represent the bit depth. If the input array is neither of them, the function will raise an error. The format of half precision floating point is defined in IEEE 754-2008.

pub fn mean(&self) -> Scalar

Calculates the mean value M of array elements, independently for each channel, and return it as Scalar. TODO: pass second paramter with mask

pub fn sum(&self) -> Scalar

Calculates the per-channel pixel sum of an image.

pub fn lookup_table_transform(&self, table: &Mat, dst: &mut Mat)

Performs a look-up table transform of an array.

The function LUT fills the output array with values from the look-up table. Indices of the entries are taken from the input array.

pub fn rows(&self) -> i32

Returns the number of rows for this Mat.

pub fn cols(&self) -> i32

Returns the number of cols for this Mat.

pub fn channels(&self) -> i32

Returns the number of channels for this Mat.

pub fn cv_type(&self) -> CvType

Returns the type for this Mat.

pub fn cv_step(&self) -> i32

Returns the number of bytes each matrix row occupies.

pub fn uchar_at(&self, row: i32, col: i32) -> u8

Returns a value from a specific row/col in this Mat (must be CV_8U).

pub fn uchar_at3(&self, x: i32, y: i32, z: i32) -> u8

Returns a value from a specific x, y, z in this Mat (must be CV_8U).

pub fn schar_at(&self, row: i32, col: i32) -> i8

Returns a value from a specific row/col in this Mat (must be CV_8S).

pub fn schar_at3(&self, x: i32, y: i32, z: i32) -> i8

Returns a value from a specific x, y, z in this Mat (must be CV_8S).

pub fn short_at(&self, row: i32, col: i32) -> i16

Returns a value from a specific row/col in this Mat (must be CV_16S).

pub fn short_at3(&self, x: i32, y: i32, z: i32) -> i16

Returns a value from a specific x, y, z in this Mat (must be CV_16S).

pub fn int_at(&self, row: i32, col: i32) -> i32

Returns a value from a specific row/col in this Mat (must be CV_32S).

pub fn int_at3(&self, x: i32, y: i32, z: i32) -> i32

Returns a value from a specific x, y, z in this Mat (must be CV_32S).

pub fn float_at(&self, row: i32, col: i32) -> f32

Returns a value from a specific row/col in this Mat (must be CV_32F).

pub fn float_at3(&self, x: i32, y: i32, z: i32) -> f32

Returns a value from a specific x, y, z in this Mat (must be CV_32F).

pub fn double_at(&self, row: i32, col: i32) -> f64

Returns a value from a specific row/col in this Mat (must be CV_64F).

pub fn double_at3(&self, x: i32, y: i32, z: i32) -> f64

Returns a value from a specific x, y, z in this Mat (must be CV_64F).

pub fn set_uchar_at(&mut self, row: i32, col: i32, val: u8)

Sets a value to a specific row/col in this Mat (must be CV_8U).

pub fn set_uchar_at3(&mut self, x: i32, y: i32, z: i32, val: u8)

Sets a value to a specific x, y, z in this Mat (must be CV_8U).

pub fn set_schar_at(&mut self, row: i32, col: i32, val: i8)

Sets a value to a specific row/col in this Mat (must be CV_8S).

pub fn set_schar_at3(&mut self, x: i32, y: i32, z: i32, val: i8)

Sets a value to a specific x, y, z in this Mat (must be CV_8S).

pub fn set_short_at(&mut self, row: i32, col: i32, val: i16)

Sets a value to a specific row/col in this Mat (must be CV_16S).

pub fn set_short_at3(&mut self, x: i32, y: i32, z: i32, val: i16)

Sets a value to a specific x, y, z in this Mat (must be CV_16S).

pub fn set_int_at(&mut self, row: i32, col: i32, val: i32)

Sets a value to a specific row/col in this Mat (must be CV_32S).

pub fn set_int_at3(&mut self, x: i32, y: i32, z: i32, val: i32)

Sets a value to a specific x, y, z in this Mat (must be CV_32S).

pub fn set_float_at(&mut self, row: i32, col: i32, val: f32)

Sets a value to a specific row/col in this Mat (must be CV_32F).

pub fn set_float_at3(&mut self, x: i32, y: i32, z: i32, val: f32)

Sets a value to a specific x, y, z in this Mat (must be CV_32F).

pub fn set_double_at(&mut self, row: i32, col: i32, val: f64)

Sets a value to a specific row/col in this Mat (must be CV_64F).

pub fn set_double_at3(&mut self, x: i32, y: i32, z: i32, val: f64)

Sets a value to a specific x, y, z in this Mat (must be CV_64F).