Struct rustcv::cuda::GpuCascade

pub struct GpuCascade { /* fields omitted */ }

Data structure that performs object detection with a cascade classifier.

Methods

impl GpuCascade

pub fn new<P: AsRef<Path>>(path: P) -> Result<Self, Error>

Creates the classifier from a file.

Name of the file from which the classifier is loaded. Only the old haar classifier (trained by the haar training application) and NVIDIA's nvbin are supported for HAAR and only new type of OpenCV XML cascade supported for LBP. The working haar models can be found at opencv_folder/data/haarcascades_cuda/.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn detect_multiscale(&self, mat: &GpuMat) -> Vec<Rect>

Detects objects of different sizes in the input image.