Struct rustcv::cuda::GpuMat

pub struct GpuMat { /* fields omitted */ }

Base storage class for GPU memory with reference counting.

Its interface matches the Mat interface with the following limitations: (1) no arbitrary dimensions support (only 2D), (2) no functions that return references to their data (because references on GPU are not valid for CPU), (3) no expression templates technique support/

Beware that the latter limitation may lead to overloaded matrix operators that cause memory allocations. The GpuMat class is convertible to cuda::PtrStepSz and cuda::PtrStep so it can be passed directly to the kernel.

Methods

impl GpuMat

pub fn new() -> GpuMat

Default constructor.

pub fn rows(&self) -> i32

Number of rows.

pub fn cols(&self) -> i32

Number of cols.

pub fn empty(&self) -> bool

Whether or not the GpuMat is empty.

pub fn upload(&mut self, mat: &Mat)

Uploads a normal Mat