Enum rustcv::CvError

pub enum CvError {
    InvalidString(String),
    InvalidPath(PathBuf),
    InvalidCascadeModel(PathBuf),
    EntryNotFound(PathBuf),
    EnumFromPrimitiveConversionError {
        value: i32,
    },
    UnknownError(String),
    UnicodeChars(String),
}

Custom errors.

Variants
InvalidString(String)

Indicates that string was invalid

InvalidPath(PathBuf)

Indicates that path was invalid

InvalidCascadeModel(PathBuf)

Indicates that cascade model was invalid

EntryNotFound(PathBuf)

Indicates that there is no entry on specified path

EnumFromPrimitiveConversionError

Indicates that conversion from primitive to enum type is failed

Fields of EnumFromPrimitiveConversionError

value: i32

Value that caused an error

UnknownError(String)

Indicates that error occurred in C++ code

UnicodeChars(String)

Indicates that string contains non ascii characters