Struct rustcv::features2d::SimpleBlobDetector

pub struct SimpleBlobDetector { /* fields omitted */ }

Class for extracting blobs from an image.

The class implements a simple algorithm for extracting blobs from an image:

Convert the source image to binary images by applying thresholding with several thresholds from minThreshold (inclusive) to maxThreshold (exclusive) with distance thresholdStep between neighboring thresholds.

Extract connected components from every binary image by findContours and calculate their centers.

Group centers from several binary images by their coordinates. Close centers form one group that corresponds to one blob, which is controlled by the minDistBetweenBlobs parameter.

From the groups, estimate final centers of blobs and their radiuses and return as locations and sizes of keypoints.

Methods

impl SimpleBlobDetector

pub fn new() -> Self

Returns a new SimpleBlobDetector.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn detect(&self, src: &Mat) -> Vec<KeyPoint>

Detect keypoints in an image using SimpleBlobDetector.