Struct rustcv::highgui::Window

pub struct Window { /* fields omitted */ }

Window is a wrapper around OpenCV's "HighGUI" named windows. While OpenCV was designed for use in full-scale applications and can be used within functionally rich UI frameworks (such as Qt*, WinForms*, or Cocoa*) or without any UI at all, sometimes there it is required to try functionality quickly and visualize the results. This is what the HighGUI module has been designed for.

Methods

impl Window

pub fn new(name: &str, flag: WindowFlag) -> Result<Self, Error>

Creates a new named window.

pub fn name(&self) -> String

Returns the window name.

pub fn is_open(&self) -> bool

Checks if the Window is open.

pub fn show(&self, image: &Mat)

Displays an image in the specified window. This function should be followed by the WaitKey function which displays the image for specified milliseconds. Otherwise, it won't display the image.

pub fn wait_key(&self, delay: i32) -> i32

Waits for a pressed key. This function is the only method in OpenCV's HighGUI that can fetch and handle events, so it needs to be called periodically for normal event processing

pub fn get_property(&self, flag: WindowProperty) -> f64

Returns properties of a window.

pub fn set_property(&mut self, flag: WindowProperty, value: WindowFlag)

Changes parameters of a window dynamically.

pub fn set_title(&mut self, title: &str) -> Result<(), Error>

Changes Window name dynamically.

pub fn move_window(&mut self, x: i32, y: i32)

Moves window to the specified position.

pub fn resize(&mut self, width: i32, height: i32)

Resizes window to the specified size.

pub fn close(&mut self)

Closes the window.