Struct rustcv::objdetect::CascadeClassifier

pub struct CascadeClassifier { /* fields omitted */ }

Cascade classifier class for object detection.

Methods

impl CascadeClassifier

pub fn new() -> Self

Creates a cascade classifier, uninitialized. Before use, call load.

pub fn load<P: AsRef<Path>>(&mut self, path: P) -> Result<(), Error>

Loads the classifier model from a path.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn detect_multiscale(&self, mat: &Mat) -> Vec<Rect>

Detects objects of different sizes in the input Mat image. The detected objects are returned as a vector of rectangles.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn detect_multiscale_with_params(
&self,
mat: &Mat,
scale: f64,
min_neighbors: i32,
min_size: Size,
max_size: Size
) -> Vec<Rect>

Detects the object using parameters specified.

mat - Matrix of the type CV_8U containing an image where objects are detected.

scale_factor - Parameter specifying how much the image size is reduced at each image scale.

min_neighbors - Parameter specifying how many neighbors each candidate rectangle should have to retain it.

min_size - Minimum possible object size. Objects smaller than that are ignored.

max_size - Maximum possible object size. Objects larger than that are ignored

OpenCV has a parameter (flags) that's not used at all.